massive overhaul
A complete overhaul of how this repo is organised. This is very much still a work in process. I am all to aware that a lot of this work will rot unless it is made more accessible. The individual models have been seperated out and instruction has been provided for including the relevant folders.

To be continued...